backend-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
aiosqlite==0.21.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==6.4.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<4
pydantic-settings==2.10.1
email-validator==2.3.0
httpx==0.28.1
openai==1.107.0
yfinance==0.2.65
pandas==2.3.2
python-multipart==0.0.20
